add cart page

+++ seatSelection/index.php
@@ -1,6 +1,7 @@
 <?php 
     include "../dbconnect.php";
     $timeslotid = $_GET['timeslotid'];
+    $timeslotid_str = strval ($timeslotid);
     $query = "select * from showtime where timeslotid=".$timeslotid;
     $result = $db->query($query);
     if (!$result) {
@@ -57,12 +58,17 @@
 
     if (isset($_GET['seat'])) {
         $seatarray = explode(",", $_GET['seat']);
+        $arr1 = [];
+        $timeslot_str = strval($timeslotid);
         foreach($seatarray as $seatid){
-            $_SESSION['cart'][] = $seatid;
+            $_SESSION['cart'][$timeslot_str][] = $seatid;
         }
         header('location: ' . $_SERVER['PHP_SELF']. '?' . SID.'&timeslotid='.$timeslotid);
         exit();
+
     }
+
+    // var_dump($_SESSION);
 ?>
 
 <!DOCTYPE <!DOCTYPE html>
@@ -97,12 +103,13 @@
             </ul>
             
 
-            <button type="submit" id="addto-cart" onclick="add_to_cart()" disabled>Add to cart</button>           
+            <button id="addto-cart" onclick="add_to_cart()" disabled>Add to cart</button>           
             <div class="confirm-cancel">
             <a href="../cart/index.php">
-                <button type="submit" id="confirm" disabled>Confirm</button>
+                <button id="confirm">Confirm</button>
             </a>
-            <button type="submit" id="cancel">Cancel</button>
+            <a href="../movies/index.php">
+            <button id="cancel">Cancel</button></a>
             </div>  
         </div>
 
@@ -147,15 +154,18 @@
                                 continue;
                             }
                             if ($_SESSION['cart']){
-                                foreach($_SESSION['cart'] as $seat_reserved){
-                                    if ($seat_reserved == chr($i+65).$j){
-                                        echo "<td><input type='checkbox' name='reserved' value='".chr($i+65).$j."' id='".chr($i+65).$j."'disabled>";
-                                        echo "<label for='".chr($i+65).$j."'>".chr($i+65).$j."</label>";
-                                        echo "</td>";
-                                        $disable = 1;
-                                        break; 
+                                if (array_key_exists($timeslotid_str,$_SESSION['cart'])){
+                                    foreach($_SESSION['cart'][$timeslotid_str] as $seat_reserved){
+                                        if ($seat_reserved == chr($i+65).$j){
+                                            echo "<td><input type='checkbox' name='reserved' value='".chr($i+65).$j."' id='".chr($i+65).$j."'disabled>";
+                                            echo "<label for='".chr($i+65).$j."'>".chr($i+65).$j."</label>";
+                                            echo "</td>";
+                                            $disable = 1;
+                                            break; 
+                                        }
                                     }
                                 }
+
                                 if ($disable == 1){
                                     $disable = 0;
                                     continue;
